functions/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appengine-python-standard==1.1.6
attrs==24.2.0
azure-core==1.32.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.9.0
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
firebase-admin==6.5.0
firebase-functions==0.1.2
Flask==3.0.3
Flask-Cors==5.0.0
frozendict==2.4.6
frozenlist==1.5.0
functions-framework==3.8.1
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-events==0.13.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
IMAPClient==3.0.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
MarkupSafe==3.0.2
microsoft-kiota-abstractions==1.6.0
microsoft-kiota-authentication-azure==1.6.0
microsoft-kiota-http==1.6.0
microsoft-kiota-serialization-form==1.6.0
microsoft-kiota-serialization-json==1.6.0
microsoft-kiota-serialization-multipart==1.6.0
microsoft-kiota-serialization-text==1.6.0
mock==5.1.0
msgpack==1.1.0
msgraph-core==1.1.6
msgraph-sdk==1.11.0
multidict==6.1.0
nodeenv==1.9.1
openai==1.54.3
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
packaging==24.1
pendulum==3.0.0
pillow==11.0.0
pip-autoremove==0.10.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMuPDF==1.24.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
std-uritemplate==1.0.6
time-machine==2.16.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
watchdog==6.0.0
Werkzeug==3.1.2
wrapt==1.16.0
yarl==1.17.1
zipp==3.20.2
